added lr stuff

+++ src/main/java/com/lr/model/Consignee.java
@@ -143,7 +143,7 @@ public class Consignee implements Serializable {
 	
 	private static final String QUERY_FOR_CONSIGNEE_BY_ID_SKEY =
 		Consignee.class.getName() + ".findConsigneeById";
-	public static Consignee findConsigneeById(Session session, Long id)
+	public static Consignee findConsigneeById(Session session, Integer id)
 		throws HibernateException
 	{
 	 	if (id == null) {
@@ -151,7 +151,7 @@ public class Consignee implements Serializable {
 	 	}
 	 	
 	 	Query qry = session.getNamedQuery(QUERY_FOR_CONSIGNEE_BY_ID_SKEY);
-	 	qry.setLong("id", id);		 	
+	 	qry.setInteger("id", id);		 	
  
 	 	qry.setMaxResults(1);
